refactor : 패키지 정리
- domain, presentation 으로 간단히 정리
- 테스트 패키지 또한 같은 기준으로 정
