+++ src/main/java/lotto/presentation/InputView.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.presentation;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
