+++ src/test/java/lotto/domain/LottoGroupTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
 import org.assertj.core.api.Assertions;
