+++ src/test/java/lotto/presentation/LottoResultFormatterTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.presentation;
 
+import lotto.domain.Prize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
